% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ma}
\alias{ma}
\title{Moving average}
\usage{
ma(x, n = 5, sides = 2)
}
\arguments{
\item{x}{A vector to create the moving average for.}

\item{n}{The number of samples to create the moving average for.}

\item{sides}{the}
}
\description{
Moving average
}
