% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpackagefromorg-package.R
\docType{package}
\name{rpackagefromorg-package}
\alias{rpackagefromorg}
\alias{rpackagefromorg-package}
\title{rpackagefromorg: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Ilja Kocken \email{i.j.kocken@uu.nl} (\href{https://orcid.org/0000-0003-2196-8718}{ORCID})

}
\keyword{internal}
